+++ android/src/main/kotlin/com/wttec/android_webview/internal/DownFileProvider.java
@@ -1,0 +1,6 @@
+package com.wttec.android_webview.internal;
+
+import android.support.v4.content.FileProvider;
+
+public class DownFileProvider extends FileProvider {
+}