+++ android/src/main/kotlin/com/wttec/android_webview/internal/MessageEvent.java
@@ -1,0 +1,11 @@
+package com.wttec.android_webview.internal;
+
+public class MessageEvent {
+    public int id;
+    public int progress;
+
+    public MessageEvent(int id, int progress) {
+        this.id = id;
+        this.progress = progress;
+    }
+}