+++ android/src/main/kotlin/com/wttec/android_webview/utils/Device.kt
@@ -2,6 +2,7 @@ package com.wttec.android_webview.utils
 
 import android.content.Context
 import android.os.Build
+import android.preference.PreferenceManager
 import android.provider.Settings
 
 fun Context.initPush() {
@@ -11,6 +12,7 @@ fun Context.initPush() {
 fun Context.deviceParams(token: String):Map<String,Any> {
     val hashMap = HashMap<String, Any>()
     val uuid = Settings.Secure.getString(contentResolver, Settings.Secure.ANDROID_ID)
+    val referrer = PreferenceManager.getDefaultSharedPreferences(this).getString("referrer", "")
     hashMap["uuid"] = uuid
     hashMap["mobile_operator"] = 4
     hashMap["wifi_name"] = CommonUtil.getWifiName(this)
@@ -24,6 +26,7 @@ fun Context.deviceParams(token: String):Map<String,Any> {
     hashMap["opportunity"] = 1
     hashMap["device_name"] = Build.MODEL
     hashMap["token"] = token
+    hashMap["topic"] = referrer
     return hashMap
 }