+++ lib/android_webview.dart
@@ -4,6 +4,7 @@ import 'package:flutter/services.dart';
 
 class AndroidWebview {
   static const MethodChannel _channel = const MethodChannel('com.wttec.android_webview');
+  static const EventChannel eventChannel = const EventChannel("com.wttec.android_webview/event");
 
   static Future<String> get platformVersion async {
     final String version = await _channel.invokeMethod('getPlatformVersion');
@@ -18,8 +19,8 @@ class AndroidWebview {
     return _channel.invokeMethod("logout");
   }
 
-  static void toWeb(String url, {int type}) {
-    _channel.invokeMethod("toWeb", {"url": url, "openType": type});
+  static void toWeb(String name, int type, int id, String url) {
+    _channel.invokeMethod("toWeb", {"url": url, "openType": type, "name": name, "id": id});
   }
 
   static Future<Map<String, dynamic>> get tokenMap async {
@@ -43,6 +44,15 @@ class AndroidWebview {
   static Future<String> get version async {
     final String result = await _channel.invokeMethod("version");
     return result;
+  }
+
+  static Future<bool> isDownloaded(int id) async {
+    final bool result = await _channel.invokeMethod("isDownloaded", id);
+    return result;
+  }
+
+  static void install(int id) {
+    _channel.invokeMethod("install", id);
   }
 
   static void setHandler(Future<dynamic> handler(MethodCall call)) {