+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 0.1.1
+
+*
+
+## 0.1.0
+
+* download apk progress
+
+
+## 0.0.7
+
+* download apk progress
+
+## 0.0.6
+
+* delete unused code.
+
+## 0.0.5
+
+* install referrer
+
 ## 0.0.4
 
 * bug fix.